# src/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IVMResponse": {
				"properties": {
					"cpuUsage": {
						"type": "string"
					},
					"ramUsagePercent": {
						"type": "string"
					},
					"ramUsage": {
						"type": "string"
					},
					"totalRam": {
						"type": "string"
					},
					"freeRam": {
						"type": "string"
					},
					"totalDisk": {
						"type": "string"
					},
					"freeDisk": {
						"type": "string"
					},
					"diskUsage": {
						"type": "string"
					},
					"diskUsagePercent": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"staticId": {
						"type": "string"
					},
					"dynamicId": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"macAddress": {
						"type": "string"
					},
					"ipAddress": {
						"type": "string"
					},
					"port": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pm2ProcessResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"pid": {
						"type": "number",
						"format": "double"
					},
					"pm_id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"cpu": {
						"type": "number",
						"format": "double"
					},
					"memory": {
						"type": "number",
						"format": "double"
					},
					"uptime": {
						"type": "number",
						"format": "double"
					},
					"cwd": {
						"type": "string"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					},
					"outLogPath": {
						"type": "string"
					},
					"errLogPath": {
						"type": "string"
					},
					"restarts": {
						"type": "number",
						"format": "double"
					},
					"staticId": {
						"type": "string"
					},
					"dynamicId": {
						"type": "string"
					},
					"heapMemory": {},
					"monit": {}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Pm2StatusSummaryResponse": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"online": {
						"type": "number",
						"format": "double"
					},
					"stopped": {
						"type": "number",
						"format": "double"
					},
					"errored": {
						"type": "number",
						"format": "double"
					},
					"other": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"online",
					"stopped",
					"errored",
					"other"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Pm2LogType": {
				"type": "string",
				"enum": [
					"out",
					"err",
					"all"
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "spinal-organ-health",
		"version": "1.0.1",
		"description": "API for Spinal Organ Health Monitoring",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "SpinalCom",
			"email": "developers@spinalcom.com"
		}
	},
	"paths": {
...
		"/monitoring/{vmKey}": {
			"get": {
				"operationId": "GetVirtualMachine",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IVMResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/{vmKey}/system": {
			"get": {
				"operationId": "GetSystemMetrics",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IVMResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"description": "Returns current CPU, memory, and system-level metrics collected by the agent.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/{vmKey}/apps": {
			"get": {
				"operationId": "GetApps",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/Pm2ProcessResponse"
											},
											"type": "array"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"description": "Lists all PM2-managed applications with their current runtime state.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/monitoring/{vmKey}/apps/status/summary": {
			"get": {
				"operationId": "GetAppsStatusSummary",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Pm2StatusSummaryResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"description": "Returns PM2 process counts grouped by status.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/monitoring/{vmKey}/apps/{key}": {
			"get": {
				"operationId": "GetAppByKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Pm2ProcessResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Process not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Returns details for a single PM2 application by name or identifier key.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "PM2 process key used to find the app.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/monitoring/{vmKey}/apps/{key}/logs": {
			"get": {
				"operationId": "GetAppLogsByKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Pm2ProcessLogsResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Process not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Returns tailed stdout/stderr logs for one PM2 application.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "PM2 process key used to find the app.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Number of lines to return per stream (1..1000).",
						"in": "query",
						"name": "tail",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Which stream to return: out, err, or all.",
						"in": "query",
						"name": "logType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Pm2LogType"
						}
					}
				]
			}
		},
		"/monitoring/{vmKey}/apps/start": {
			"post": {
				"operationId": "StartApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"failed": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												},
												"started": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												}
											},
											"required": [
												"failed",
												"started"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Unable to start process",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Starts one or more PM2 applications.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Request body containing the list of application keys to start.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"keys": {
										"items": {
											"anyOf": [
												{
													"type": "string"
												},
												{
													"type": "number",
													"format": "double"
												}
											]
										},
										"type": "array"
									}
								},
								"required": [
									"keys"
								],
								"type": "object",
								"description": "Request body containing the list of application keys to start."
							}
						}
					}
				}
			}
		},
		"/monitoring/{vmKey}/apps/stop": {
			"post": {
				"operationId": "StopApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"failed": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												},
												"stopped": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												}
											},
											"required": [
												"failed",
												"stopped"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Unable to stop process",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Stops one or more PM2 applications.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Request body containing the list of application keys to stop.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"keys": {
										"items": {
											"anyOf": [
												{
													"type": "string"
												},
												{
													"type": "number",
													"format": "double"
												}
											]
										},
										"type": "array"
									}
								},
								"required": [
									"keys"
								],
								"type": "object",
								"description": "Request body containing the list of application keys to stop."
							}
						}
					}
				}
			}
		},
		"/monitoring/{vmKey}/apps/restart": {
			"post": {
				"operationId": "RestartApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"failed": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												},
												"restarted": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												}
											},
											"required": [
												"failed",
												"restarted"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Unable to restart process",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Restarts one or more PM2 applications.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Request body containing the list of application keys to restart.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"keys": {
										"items": {
											"anyOf": [
												{
													"type": "string"
												},
												{
													"type": "number",
													"format": "double"
												}
											]
										},
										"type": "array"
									}
								},
								"required": [
									"keys"
								],
								"type": "object",
								"description": "Request body containing the list of application keys to restart."
							}
						}
					}
				}
			}
		},
		"/monitoring/{vmKey}/apps/reload": {
			"post": {
				"operationId": "ReloadApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"failed": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												},
												"reloaded": {
													"items": {
														"$ref": "#/components/schemas/ActionResponse"
													},
													"type": "array"
												}
											},
											"required": [
												"failed",
												"reloaded"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Unable to reload process",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Reloads one or more PM2 applications.",
				"tags": [
					"Monitoring"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vmKey",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Request body containing the list of application keys to reload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"keys": {
										"items": {
											"anyOf": [
												{
													"type": "string"
												},
												{
													"type": "number",
													"format": "double"
												}
											]
										},
										"type": "array"
									}
								},
								"required": [
									"keys"
								],
								"type": "object",
								"description": "Request body containing the list of application keys to reload."
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}